refactor(web): extract `toc` and `lastReadChapter` computed properties into separate hooks
