+++ web/src/pages/novel/components/WebNovelWide.vue
@@ -3,8 +3,7 @@ import { SortOutlined } from '@vicons/material';
 
 import { Locator } from '@/data';
 import { WebNovelTocItemDto, WebNovelDto } from '@/model/WebNovel';
-
-import { ReadableTocItem } from './common';
+import { useToc, useLastReadChapter } from '@/hooks/useWebNovel';
 
 const props = defineProps<{
   providerId: string;
@@ -14,24 +13,8 @@ const props = defineProps<{
 
 const { setting } = Locator.settingRepository();
 
-const toc = computed(() => {
-  const { novel } = props;
-  const novelToc = novel.toc as ReadableTocItem[];
-  let order = 0;
-  for (const [index, it] of novelToc.entries()) {
-    it.key = index;
-    it.order = it.chapterId ? order : undefined;
-    if (it.chapterId) order += 1;
-  }
-  return novelToc;
-});
-
-const lastReadChapter = computed(() => {
-  const { novel } = props;
-  if (novel.lastReadChapterId) {
-    return toc.value.find((it) => it.chapterId === novel.lastReadChapterId);
-  }
-});
+const { toc } = useToc(props.novel);
+const { lastReadChapter } = useLastReadChapter(props.novel, toc);
 </script>
 
 <template>
